Convert List directory
