--- src/frontend/sharedComponents/List/ListContext.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const ListContext = React.createContext({});
-
-if (process.env.NODE_ENV !== "production") {
-  ListContext.displayName = "ListContext";
-}
-
-export default ListContext;
