+++ src/frontend/sharedComponents/List/ListContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+export const ListContext = createContext<{ dense?: boolean }>({});
+
+if (process.env.NODE_ENV !== "production") {
+  ListContext.displayName = "ListContext";
+}
