--- src/frontend/sharedComponents/List/index.js
@@ -1,4 +1,0 @@
-export { default as List } from "./List";
-export { default as ListItem } from "./ListItem";
-export { default as ListItemText } from "./ListItemText";
-export { default as ListItemIcon } from "./ListItemIcon";
