+++ src/frontend/sharedComponents/List/index.ts
@@ -1,0 +1,4 @@
+export * from "./List";
+export * from "./ListItem";
+export * from "./ListItemText";
+export * from "./ListItemIcon";
